+++ lmcache/v1/distributed/l2_adapters/nixl_store_dynamic_l2_adapter.py
@@ -24,20 +24,9 @@ Secondary lookup (always on):
 from __future__ import annotations
 
 # Standard
-from abc import ABC, abstractmethod
-from typing import Any, Optional
+from typing import Optional
 import asyncio
-import os
 import threading
-import uuid
-
-# Third Party
-from nixl._api import nixl_agent as NixlAgent
-from nixl._api import nixl_agent_config as NixlAgentConfig
-from nixl._api import nixl_xfer_handle as NixlXferHandle
-from nixl._api import (
-    nixlBind,
-)
 
 # First Party
 from lmcache.lmcache_native import Bitmap
@@ -52,6 +41,13 @@ from lmcache.v1.distributed.l2_adapters.config import (
 from lmcache.v1.distributed.l2_adapters.factory import (
     register_l2_adapter_factory,
 )
+from lmcache.v1.distributed.l2_adapters.nixl_store_agents.dynamic_nixl_store_agent import (  # noqa: E501
+    DynamicNixlStorageAgent,
+)
+from lmcache.v1.distributed.l2_adapters.nixl_store_agents.file_dynamic_nixl_store_agent import (  # noqa: E501
+    FILE_DYNAMIC_BACKENDS,
+    FileDynamicNixlStorageAgent,
+)
 from lmcache.v1.distributed.l2_adapters.nixl_store_l2_adapter import (
     NixlStoreObj,
 )
@@ -59,378 +55,6 @@ from lmcache.v1.memory_management import MemoryObj
 from lmcache.v1.platform import create_event_notifier
 
 logger = init_logger(__name__)
-
-_FILE_DYNAMIC_BACKENDS = ("GDS", "GDS_MT", "POSIX", "HF3FS")
-
-# ---------------------------------------------------------------
-# ObjectKey <-> file path helpers
-# ---------------------------------------------------------------
-
-
-def _object_key_to_filename(key: ObjectKey) -> str:
-    """Derive a deterministic file name from an ObjectKey.
-
-    Replaces ``/`` in model names with ``--`` to avoid creating
-    subdirectories (e.g. ``meta-llama/Llama-3-8B`` becomes
-    ``meta-llama--Llama-3-8B``).
-    """
-    safe_model_name = key.model_name.replace("/", "--")
-    chunk_hex = key.chunk_hash.hex()
-    return (
-        f"{safe_model_name}_{key.kv_rank:08x}_{key.object_group_id:x}_{chunk_hex}.bin"
-    )
-
-
-def _object_key_to_relpath(key: ObjectKey) -> str:
-    """Relative path ``<hex[:2]>/<hex[2:4]>/filename`` — a 2-level hash-prefix
-    subdir tree (GDS-style) keyed on the chunk-hash hex.
-
-    ``hex`` is ``chunk_hash.hex()``, the same value embedded in the filename, so
-    the two subdir levels are the first four hex chars of the hash and match the
-    filename's hash prefix (e.g. ``834e...`` -> ``83/4e/``). Spreads files
-    across up to 256*256 subdirectories instead of one flat directory.
-    """
-    h = key.chunk_hash.hex()
-    return os.path.join(h[:2], h[2:4], _object_key_to_filename(key))
-
-
-# ---------------------------------------------------------------
-# Dynamic Nixl storage agent
-# ---------------------------------------------------------------
-
-
-class DynamicNixlStorageAgent(ABC):
-    """Base class for dynamic NIXL storage agents.
-
-    This class owns the NIXL agent, the L1 memory registration, and the common
-    transfer lifecycle. Concrete agents map :class:`ObjectKey` values to their
-    storage backend without exposing backend-specific paths to callers.
-    """
-
-    def __init__(
-        self,
-        device: str,
-        backend: str,
-        backend_params: dict[str, str],
-        l1_memory_desc: L1MemoryDesc,
-    ) -> None:
-        self.backend = backend
-        self.device = device
-        self.backend_params = backend_params
-        self.l1_align_bytes = l1_memory_desc.align_bytes
-
-        self.agent_name = "DynNixlAgent_" + str(uuid.uuid4())
-        nixl_conf = NixlAgentConfig(backends=[])
-        self.nixl_agent = NixlAgent(self.agent_name, nixl_conf)
-        self.nixl_agent.create_backend(backend, backend_params)
-
-        self._init_mem_handlers(
-            device,
-            l1_memory_desc.ptr,
-            l1_memory_desc.size,
-            l1_memory_desc.align_bytes,
-            device_id=0,
-        )
-
-    def _init_mem_handlers(
-        self,
-        device: str,
-        buffer_ptr: int,
-        buffer_size: int,
-        page_size: int,
-        device_id: int,
-    ) -> None:
-        """Register the L1 memory used by all dynamic storage transfers."""
-        reg_list = [(buffer_ptr, buffer_size, device_id, "")]
-        xfer_desc = [
-            (base_addr, page_size, device_id)
-            for base_addr in range(buffer_ptr, buffer_ptr + buffer_size, page_size)
-        ]
-
-        mem_type = "DRAM" if device == "cpu" else "VRAM"
-
-        self.mem_reg_descs = self.nixl_agent.register_memory(
-            reg_list, mem_type=mem_type
-        )
-        xfer_descs = self.nixl_agent.get_xfer_descs(xfer_desc, mem_type=mem_type)
-        self.mem_xfer_handler = self.nixl_agent.prep_xfer_dlist(
-            "", xfer_descs, mem_type=mem_type
-        )
-
-    async def _transfer(
-        self,
-        direction: str,
-        mem_indices: list[int],
-        storage_xfer_handler: Any,
-    ) -> None:
-        """Transfer pages between L1 memory and a prepared storage handler."""
-        storage_indices = list(range(len(mem_indices)))
-        handle = self.nixl_agent.make_prepped_xfer(
-            direction,
-            self.mem_xfer_handler,
-            mem_indices,
-            storage_xfer_handler,
-            storage_indices,
-        )
-        try:
-            await self._post_non_blocking(handle)
-        finally:
-            self.nixl_agent.release_xfer_handle(handle)
-
-    async def _post_non_blocking(self, handle: Any) -> None:
-        """Await a NIXL transfer until it completes or fails."""
-        state = self.nixl_agent.transfer(handle)
-        while state != "DONE" and state != "ERR":
-            try:
-                state = self.nixl_agent.check_xfer_state(handle)
-            except nixlBind.nixlBackendError:
-                raise
-            await asyncio.sleep(0.01)
-        if state == "ERR":
-            raise RuntimeError("NIXL transfer failed")
-
-    def get_memory_indices(self, raw_addr: int, mem_size: int) -> list[int]:
-        """Return the registered L1 page indices for a memory range."""
-        if raw_addr % self.l1_align_bytes != 0:
-            raise ValueError(
-                f"Raw address {raw_addr} is not aligned to "
-                f"page size {self.l1_align_bytes}"
-            )
-        if mem_size % self.l1_align_bytes != 0:
-            raise ValueError(
-                f"Memory size {mem_size} is not a multiple of "
-                f"page size {self.l1_align_bytes}"
-            )
-        num_pages = mem_size // self.l1_align_bytes
-        return [(raw_addr // self.l1_align_bytes + i) for i in range(num_pages)]
-
-    def close(self) -> None:
-        """Release resources owned by this storage agent."""
-        self.nixl_agent.release_dlist_handle(self.mem_xfer_handler)
-        self.nixl_agent.deregister_memory(self.mem_reg_descs)
-
-    @abstractmethod
-    async def dynamic_store(self, mem_indices: list[int], key: ObjectKey) -> None:
-        """Store L1 memory pages for ``key`` in the backing storage."""
-
-    @abstractmethod
-    async def dynamic_load(self, mem_indices: list[int], key: ObjectKey) -> None:
-        """Load the stored pages for ``key`` into L1 memory."""
-
-    @abstractmethod
-    def dynamic_delete(self, key: ObjectKey) -> None:
-        """Delete the backing-storage object associated with ``key``."""
-
-    @abstractmethod
-    def get_stored_size(self, key: ObjectKey) -> int | None:
-        """Return the stored size for ``key``, or ``None`` if it is absent."""
-
-    @abstractmethod
-    def cleanup(self) -> None:
-        """Perform best-effort cleanup of backend-specific transient data."""
-
-
-class FileDynamicNixlStorageAgent(DynamicNixlStorageAgent):
-    """Dynamic NIXL storage agent backed by one file per object.
-
-    The L1 memory handler is registered once at initialization. Each file is
-    registered for the duration of an individual store or load transfer and
-    is deregistered immediately afterward.
-    """
-
-    def __init__(
-        self,
-        device: str,
-        backend: str,
-        backend_params: dict[str, str],
-        l1_memory_desc: L1MemoryDesc,
-    ) -> None:
-        super().__init__(device, backend, backend_params, l1_memory_desc)
-        self.file_path = backend_params["file_path"]
-        os.makedirs(self.file_path, exist_ok=True)
-        self.use_direct_io = (
-            str(backend_params.get("use_direct_io", "false")).lower() == "true"
-        )
-        # Opt-in: spread per-key files across a fixed 2-level subdir tree
-        # instead of one flat directory. Default false = the original flat
-        # layout (unchanged). See _object_key_to_relpath / get_file_path_for_key.
-        self.shard_dirs = (
-            str(backend_params.get("shard_dirs", "false")).lower() == "true"
-        )
-        # Subdirs already created (shard_dirs only), to skip redundant makedirs
-        # on the store hot path. Bounded by the fanout (<= 256*256 entries).
-        self._created_subdirs: set[str] = set()
-
-    async def dynamic_store(self, mem_indices: list[int], key: ObjectKey) -> None:
-        """Write-to-temp-then-rename to publish the final file atomically.
-
-        The DMA write goes to ``<file_path>.tmp.<uuid>`` in the same
-        directory. Only after the transfer completes successfully is the
-        temp file atomically renamed to the final path, ensuring that
-        concurrent readers (including other processes sharing the same
-        directory) never observe a partially-written file.
-        """
-        file_path = self._get_file_path_for_key(key)
-        page_size = self.l1_align_bytes
-        file_size = len(mem_indices) * page_size
-        tmp_path = f"{file_path}.tmp.{uuid.uuid4().hex}"
-        # With sharding, create the subdir once per bucket (cached). Flat layout
-        # needs nothing here — the base dir is created at init.
-        if self.shard_dirs:
-            subdir = os.path.dirname(tmp_path)
-            if subdir not in self._created_subdirs:
-                os.makedirs(subdir, exist_ok=True)
-                self._created_subdirs.add(subdir)
-        fd = os.open(tmp_path, self._open_flags(create=True))
-        try:
-            reg_descs, xfer_handler = self._register_single_file(
-                fd, file_size, page_size
-            )
-            try:
-                await self._transfer("WRITE", mem_indices, xfer_handler)
-            finally:
-                self._deregister_file(reg_descs, xfer_handler)
-        except BaseException:
-            # Best-effort cleanup of the temp file on failure.
-            try:
-                os.unlink(tmp_path)
-            except FileNotFoundError:
-                pass
-            raise
-        finally:
-            os.close(fd)
-
-        # Atomic publish: readers only ever see a complete file at file_path.
-        # TODO(Jiayi): Only guaranteed to be atomic within the local posix filesystems.
-        os.rename(tmp_path, file_path)
-
-    async def dynamic_load(self, mem_indices: list[int], key: ObjectKey) -> None:
-        """Open an existing file, DMA read into L1 memory, then clean up."""
-        file_path = self._get_file_path_for_key(key)
-        page_size = self.l1_align_bytes
-        file_size = len(mem_indices) * page_size
-        fd = os.open(file_path, self._open_flags(create=False))
-        try:
-            reg_descs, xfer_handler = self._register_single_file(
-                fd, file_size, page_size
-            )
-            try:
-                await self._transfer("READ", mem_indices, xfer_handler)
-            finally:
-                self._deregister_file(reg_descs, xfer_handler)
-        finally:
-            os.close(fd)
-
-    def dynamic_delete(self, key: ObjectKey) -> None:
-        """Delete a storage file from disk."""
-        file_path = self._get_file_path_for_key(key)
-        try:
-            os.unlink(file_path)
-        except FileNotFoundError:
-            logger.warning("File already deleted: %s", file_path)
-
-    def get_stored_size(self, key: ObjectKey) -> int | None:
-        """Return the data-file size for ``key``, or ``None`` if it is absent."""
-        try:
-            return os.stat(self._get_file_path_for_key(key)).st_size
-        except FileNotFoundError:
-            return None
-
-    def get_memory_indices(self, raw_addr: int, mem_size: int) -> list[int]:
-        """Get L1 memory page indices for the given address and size."""
-        if raw_addr % self.l1_align_bytes != 0:
-            raise ValueError(
-                f"Raw address {raw_addr} is not aligned to "
-                f"page size {self.l1_align_bytes}"
-            )
-        if mem_size % self.l1_align_bytes != 0:
-            raise ValueError(
-                f"Memory size {mem_size} is not a multiple of "
-                f"page size {self.l1_align_bytes}"
-            )
-        num_pages = mem_size // self.l1_align_bytes
-        return [(raw_addr // self.l1_align_bytes + i) for i in range(num_pages)]
-
-    def get_file_path_for_key(self, key: ObjectKey) -> str:
-        """Return the on-disk path for ``key``: sharded when ``shard_dirs`` is
-        set (see ``_object_key_to_relpath``), otherwise the flat layout.
-        """
-        if self.shard_dirs:
-            return os.path.join(self.file_path, _object_key_to_relpath(key))
-        return os.path.join(self.file_path, _object_key_to_filename(key))
-
-    async def _post_non_blocking(self, handle: NixlXferHandle) -> None:
-        """Await a nixl transfer until done."""
-        state = self.nixl_agent.transfer(handle)
-        while state != "DONE" and state != "ERR":
-            try:
-                state = self.nixl_agent.check_xfer_state(handle)
-            except nixlBind.nixlBackendError:
-                raise
-            if state != "DONE" and state != "ERR":
-                await asyncio.sleep(0.01)
-        if state == "ERR":
-            raise RuntimeError("NIXL transfer failed")
-
-    def cleanup(self) -> None:
-        """Remove leftover ``*.tmp.*`` files in the storage directory.
-
-        These can be left behind if a store crashed between opening the
-        temp file and the atomic rename. Called at shutdown as a best-effort
-        GC; orphans don't affect correctness because they're never matched
-        by the deterministic ``ObjectKey → filename`` mapping. Walks
-        subdirectories so it also reaches temp files under the sharded layout.
-        """
-        for root, _dirs, files in os.walk(self.file_path):
-            for name in files:
-                # Temp suffix format: "<final_name>.tmp.<hex>"
-                if ".tmp." in name:
-                    try:
-                        os.unlink(os.path.join(root, name))
-                    except FileNotFoundError:
-                        pass
-                    except OSError as e:
-                        logger.warning(
-                            "Failed to remove leftover temp file %s: %s", name, e
-                        )
-
-    def _open_flags(self, create: bool) -> int:
-        """Return os.open flags for storage files."""
-        flags = os.O_RDWR
-        if create:
-            # O_TRUNC ensures any orphaned file from a previous crash
-            # is truncated, avoiding stale trailing bytes on disk.
-            flags |= os.O_CREAT | os.O_TRUNC
-        if self.use_direct_io and hasattr(os, "O_DIRECT"):
-            flags |= os.O_DIRECT
-        return flags
-
-    def _register_single_file(
-        self, fd: int, file_size: int, page_size: int
-    ) -> tuple[Any, Any]:
-        """Register a single file with NIXL for a transfer."""
-        num_pages = file_size // page_size
-        reg_list = [(0, file_size, fd, "")]
-        xfer_desc = [(offset * page_size, page_size, fd) for offset in range(num_pages)]
-        reg_descs = self.nixl_agent.register_memory(reg_list, mem_type="FILE")
-        xfer_descs = self.nixl_agent.get_xfer_descs(xfer_desc, mem_type="FILE")
-        xfer_handler = self.nixl_agent.prep_xfer_dlist(
-            self.agent_name, xfer_descs, mem_type="FILE"
-        )
-        return reg_descs, xfer_handler
-
-    def _deregister_file(self, reg_descs: Any, xfer_handler: Any) -> None:
-        """Release NIXL resources registered for a single file."""
-        self.nixl_agent.release_dlist_handle(xfer_handler)
-        self.nixl_agent.deregister_memory(reg_descs)
-
-    def _get_file_path_for_key(self, key: ObjectKey) -> str:
-        """Return the on-disk path for ``key``: sharded when ``shard_dirs`` is
-        set (see ``_object_key_to_relpath``), otherwise the flat layout.
-        """
-        if self.shard_dirs:
-            return os.path.join(self.file_path, _object_key_to_relpath(key))
-        return os.path.join(self.file_path, _object_key_to_filename(key))
 
 
 def _create_dynamic_nixl_storage_agent(
@@ -441,10 +65,19 @@ def _create_dynamic_nixl_storage_agent(
 ) -> DynamicNixlStorageAgent:
     """Create the dynamic storage agent registered for ``backend``.
 
-    The factory selects an implementation by backend family so the L2 adapter
-    does not depend on a concrete storage representation.
+    Args:
+        device: Device that owns the registered L1 memory.
+        backend: NIXL storage backend name.
+        backend_params: Backend-specific NIXL parameters.
+        l1_memory_desc: L1 memory region shared with the agent.
+
+    Returns:
+        The concrete storage agent for the requested backend.
+
+    Raises:
+        ValueError: If the dynamic adapter has no agent for ``backend``.
     """
-    if backend in _FILE_DYNAMIC_BACKENDS:
+    if backend in FILE_DYNAMIC_BACKENDS:
         return FileDynamicNixlStorageAgent(
             device, backend, backend_params, l1_memory_desc
         )
@@ -917,7 +550,7 @@ class DynamicNixlStoreL2Adapter(L2AdapterInterface):
 
 # TODO(Jiayi): OBJ backend is not supported in the dynamic adapter yet.
 # Only file-based backends are supported.
-_VALID_DYNAMIC_BACKENDS = _FILE_DYNAMIC_BACKENDS
+_VALID_DYNAMIC_BACKENDS = FILE_DYNAMIC_BACKENDS
 
 
 class DynamicNixlStoreL2AdapterConfig(L2AdapterConfigBase):
