+++ tests/v1/distributed/test_nixl_store_dynamic_l2_adapter.py
@@ -27,13 +27,17 @@ from lmcache.v1.distributed.internal_api import (  # noqa: E402
     L2AdapterListener,
 )
 from lmcache.v1.distributed.l2_adapters.config import PersistConfig  # noqa: E402
-from lmcache.v1.distributed.l2_adapters.nixl_store_dynamic_l2_adapter import (  # noqa: E402
+from lmcache.v1.distributed.l2_adapters.nixl_store_agents.dynamic_nixl_store_agent import (  # noqa: E402, E501
     DynamicNixlStorageAgent,
-    DynamicNixlStoreL2Adapter,
-    DynamicNixlStoreL2AdapterConfig,
-    FileDynamicNixlStorageAgent,
     _object_key_to_filename,
     _object_key_to_relpath,
+)
+from lmcache.v1.distributed.l2_adapters.nixl_store_agents.file_dynamic_nixl_store_agent import (  # noqa: E402, E501
+    FileDynamicNixlStorageAgent,
+)
+from lmcache.v1.distributed.l2_adapters.nixl_store_dynamic_l2_adapter import (  # noqa: E402, E501
+    DynamicNixlStoreL2Adapter,
+    DynamicNixlStoreL2AdapterConfig,
 )
 from lmcache.v1.memory_management import (  # noqa: E402
     MemoryFormat,
